Install related dependencies
Added a new `ux-install-ci` script which attempts to install dependencies with the same version as the build.

This means that, if you use the same branch name in each repo, you can have a CI build of ux-aspects-micro-focus incorporating changes in UXAspects and/or quantum-ux-aspects without having to merge first.
