+++ package.json
@@ -1,8 +1,9 @@
 {
   "name": "@ux-aspects/ux-aspects-scripts",
-  "version": "2.3.3",
+  "version": "2.4.0",
   "description": "Build scripts for UX Aspects distributables.",
   "bin": {
+    "ux-install-ci": "./bin/install-ci.js",
     "ux-install-snapshot": "./bin/install-snapshot.js",
     "ux-package-artifactory": "./bin/package-artifactory.js",
     "ux-setversion": "./bin/setversion.js",
